* fix compile errors and warnings

+++ spine-c/src/spine/SkeletonJson.c
@@ -564,7 +564,7 @@ spSkeletonData* spSkeletonJson_readSkeletonData (spSkeletonJson* self, const cha
 							mesh->a = toColor(color, 3);
 						}
 
-						mesh->hullLength = Json_getFloat(attachmentMap, "hull", 0);
+						mesh->hullLength = (int)Json_getFloat(attachmentMap, "hull", 0);
 
 						entry = Json_getItem(attachmentMap, "edges");
 						if (entry) {
@@ -635,7 +635,7 @@ spSkeletonData* spSkeletonJson_readSkeletonData (spSkeletonJson* self, const cha
 							mesh->a = toColor(color, 3);
 						}
 
-						mesh->hullLength = Json_getFloat(attachmentMap, "hull", 0);
+						mesh->hullLength = (int)Json_getFloat(attachmentMap, "hull", 0);
 
 						entry = Json_getItem(attachmentMap, "edges");
 						if (entry) {
